Re-applied the last couple of James' commits to the re-factored code base.

+++ GPConnect.Provider.AcceptanceTests/Steps/HttpSteps.cs
@@ -56,7 +56,7 @@ namespace GPConnect.Provider.AcceptanceTests.Steps
         [Then(@"the response status code should be ""(.*)""")]
         public void ThenTheResponseStatusCodeShouldBe(string statusCode)
         {
-            HttpContext.ResponseStatusCode.ToString().ShouldBe(statusCode);
+            ((int)HttpContext.ResponseStatusCode).ToString().ShouldBe(statusCode);
             Log.WriteLine("Response HttpStatusCode should be {0} but was {1}", statusCode, HttpContext.ResponseStatusCode);
         }
